#000: Poorva: Move LocationAkkaSerializer to csw-location-api jvm module.
Modules dependent on location-client or location-api need serializer on path.
