+++ build.sbt
@@ -167,6 +167,7 @@ lazy val `csw-location-api` = crossProject(JSPlatform, JVMPlatform)
   .crossType(CrossType.Full)
   .in(file("csw-location/csw-location-api"))
   .enablePlugins(PublishBintray)
+  .jvmConfigure(_.dependsOn(`csw-commons`))
   .dependsOn(`csw-prefix`)
   .jvmConfigure(_.enablePlugins(GenJavadocPlugin))
   .jvmConfigure(_.dependsOn(`csw-logging-client`).enablePlugins(MaybeCoverage))
