+++ csw-location/csw-location-api/jvm/src/main/resources/reference.conf
@@ -1,0 +1,8 @@
+akka.actor {
+  serializers {
+    location-cbor = "csw.location.api.serializer.LocationAkkaSerializer"
+  }
+  serialization-bindings {
+    "csw.location.api.codec.LocationSerializable" = location-cbor
+  }
+}
