+++ csw-location/csw-location-server/src/main/resources/application.conf
@@ -8,15 +8,6 @@ csw-location-server {
 }
 
 akka {
-  actor {
-    serializers {
-      location-cbor = "csw.location.server.internal.LocationAkkaSerializer"
-    }
-    serialization-bindings {
-      "csw.location.api.codec.LocationSerializable" = location-cbor
-    }
-  }
-
   remote {
     artery {
       enabled = on
